Разметка перенесена в JS, обработчики событий в рендер
